+++ RTracer/src/core/color.h
@@ -11,11 +11,11 @@ struct color : vec3
 	{
 	}
 
-	color(double x, double y, double z) : vec3(x, y, z)
+	color(float x, float y, float z) : vec3(x, y, z)
 	{
 	}
 
-	explicit color(double xyz) : vec3(xyz)
+	explicit color(float xyz) : vec3(xyz)
 	{
 	}
 
@@ -27,10 +27,10 @@ struct color : vec3
 	static color blue() { return color(0, 0, 1); }
 	static color green() { return color(0, 1, 0); }
 	static color magenta() { return color(1, 0, 1); }
-	static color gray() { return color(0.5, 0.5, 0.5); }
+	static color gray() { return color(0.5f, 0.5f, 0.5f); }
 };
 
-inline unsigned char to_writable_color(double rgb, double inv_samples_per_pixel)
+inline unsigned char to_writable_color(float rgb, float inv_samples_per_pixel)
 {
-	return static_cast<unsigned char>((clamp(255.0 * sqrt(inv_samples_per_pixel * rgb), 0.0, 255.0)));
+	return static_cast<unsigned char>((clamp(255.0f * sqrt(inv_samples_per_pixel * rgb), 0.0f, 255.0f)));
 }
