+++ RTracer/src/core/vec3_utility.h
@@ -12,22 +12,22 @@ inline vec3 operator-(vec3 v1, const vec3& v2)
 	return v1 += -v2;
 }
 
-inline vec3 operator*(vec3 v1, double v)
+inline vec3 operator*(vec3 v1, float v)
 {
 	return v1 *= v;
 }
 
-inline vec3 operator/(vec3 v1, double v)
+inline vec3 operator/(vec3 v1, float v)
 {
 	return v1 /= v;
 }
 
-inline vec3 operator*(double v, vec3 v1)
+inline vec3 operator*(float v, vec3 v1)
 {
 	return v1 *= v;
 }
 
-inline vec3 operator/(double v, vec3 v1)
+inline vec3 operator/(float v, vec3 v1)
 {
 	return v1 /= v;
 }
@@ -37,7 +37,7 @@ inline std::ostream& operator<<(std::ostream& out, const vec3& v)
 	return out << v.x() << ' ' << v.y() << ' ' << v.z();
 }
 
-inline double dot(const vec3& u, const vec3& v)
+inline float dot(const vec3& u, const vec3& v)
 {
 	return u.x() * v.x() + u.y() * v.y() + u.z() * v.z();
 }
@@ -66,15 +66,15 @@ inline vec3 reflect(const vec3& v, const vec3& n)
 	return v - 2 * dot(v, n) * n;
 }
 
-inline vec3 refract(const vec3& uv, const vec3& n, double refraction_ratio)
+inline vec3 refract(const vec3& uv, const vec3& n, float refraction_ratio)
 {
-	const auto cos_theta = fmin(dot(-uv, n), 1.0);
+	const auto cos_theta = fmin(dot(-uv, n), 1.0f);
 	const vec3 r_out_perp = refraction_ratio * (uv + cos_theta * n);
-	const vec3 r_out_parallel = -sqrt(fabs(1.0 - r_out_perp.length_squared())) * n;
+	const vec3 r_out_parallel = -sqrtf(fabs(1.0f - r_out_perp.length_squared())) * n;
 	return r_out_perp + r_out_parallel;
 }
 
-inline double sum(const vec3& v)
+inline float sum(const vec3& v)
 {
 	return v.x() + v.y() + v.z();
 }
