+++ RTracer/src/geometry/sphere.h
@@ -10,28 +10,28 @@
 class sphere : public hittable
 {
 public:
-	sphere(const char* name, const point3& center, double radius)
+	sphere(const char* name, const point3& center, float radius)
 		: hittable(name), center(center), radius(radius)
 	{
 	}
 
 
-	bool hit(const ray& ray, double t_min, double t_max, hit_info& info) override
+	bool hit(const ray& ray, float t_min, float t_max, hit_info& info) override
 	{
 		return hit(*this, ray, t_min, t_max, info);
 	}
 
-	static inline bool hit(sphere& sphere, const ray& ray, double t_min, double t_max, hit_info& info)
+	static inline bool hit(sphere& sphere, const ray& ray, float t_min, float t_max, hit_info& info)
 	{
 		vec3 oc = ray.origin - sphere.center;
-		double half_b = dot(ray.direction, oc);
-		double c = oc.length_squared() - sphere.radius * sphere.radius;
-		double squared_discriminant = half_b * half_b - c;
+		float half_b = dot(ray.direction, oc);
+		float c = oc.length_squared() - sphere.radius * sphere.radius;
+		float squared_discriminant = half_b * half_b - c;
 		bool is_hit = squared_discriminant >= 0;
 		if (is_hit)
 		{
-			double discriminant = std::sqrt(squared_discriminant);
-			double root = (-half_b - discriminant);
+			float discriminant = std::sqrt(squared_discriminant);
+			float root = (-half_b - discriminant);
 			is_hit = root >= t_min && root <= t_max;
 			if (!is_hit)
 			{
@@ -62,7 +62,7 @@ public:
 	}
 
 	point3 center;
-	double radius;
+	float radius;
 
 private:
 	static void set_uv_at(const vec3& p, vec3& uv_coordinates)
@@ -74,8 +74,8 @@ private:
 		//     <0 1 0> yields <0.50 1.00>       < 0 -1  0> yields <0.50 0.00>
 		//     <0 0 1> yields <0.25 0.50>       < 0  0 -1> yields <0.75 0.50>
 
-		const double theta = acos(-p.y());
-		const double phi = atan2(-p.z(), p.x()) + constants::pi;
+		const float theta = acos(-p.y());
+		const float phi = atan2(-p.z(), p.x()) + constants::pi;
 		uv_coordinates.x() = phi / (2 * constants::pi);
 		uv_coordinates.y() = theta * constants::inv_pi;
 	}
