+++ RTracer/src/raytrace_renderer.h
@@ -28,11 +28,11 @@ struct raytrace_settings
 struct raytrace_pixel
 {
 	const long index;
-	const double x;
-	const double y;
+	const float x;
+	const float y;
 	color color = color::black();
 
-	raytrace_pixel(int index, double x, double y) : index(index), x(x), y(y)
+	raytrace_pixel(int index, float x, float y) : index(index), x(x), y(y)
 	{
 	}
 };
@@ -43,8 +43,8 @@ struct raytrace_render_settings
 	color bounce_depth_limit_color = color::black();
 
 	color background_bottom_color = color::white();
-	color background_top_color = color(0.5, 0.7, 1.0);
-	double background_strength = 1.0;
+	color background_top_color = color(0.5f, 0.7f, 1.0f);
+	float background_strength = 1.0f;
 };
 
 /// <summary>
@@ -61,7 +61,7 @@ struct raytrace_render_data
 {
 	std::vector<raytrace_pixel> pixels;
 	std::vector<unsigned char> colors;
-	double iteration = 1.0;
+	float iteration = 1.0f;
 	raytrace_render_settings settings;
 
 	void set_pixels_from(const raytrace_render_data& src)
@@ -85,9 +85,9 @@ public:
 		current_render.colors.reserve(pixel_count * channels_num);
 
 		int index = 0;
-		for (int y = image_height - 1; y >= 0; --y)
+		for (float y = static_cast<float>(image_height - 1); y >= 0; --y)
 		{
-			for (int x = 0; x < image_width; ++x)
+			for (float x = 0; x < static_cast<float>(image_width); ++x)
 			{
 				current_render.pixels.emplace_back(index, x, y);
 				empty_render.pixels.emplace_back(index, x, y);
@@ -106,7 +106,7 @@ public:
 	void signal_scene_change()
 	{
 		current_render.set_pixels_from(empty_render);
-		current_render.iteration = 1.0;
+		current_render.iteration = 1.0f;
 	}
 
 	/// <summary>
@@ -126,14 +126,14 @@ public:
 	void render(const camera& camera, world& world, raytrace_render_data& data) const
 	{
 		// render settings
-		const double inv_samples_per_pixel = 1.0 / data.iteration;
+		const float inv_samples_per_pixel = 1.0f / data.iteration;
 		data.iteration++;
 
 		const raytrace_settings& raytrace_settings = this->settings;
 		const raytrace_render_settings render_settings = data.settings;
 		std::vector<unsigned char>& pixel_colors = data.colors;
-		double inv_width = 1.0 / (static_cast<double>(raytrace_settings.image_width) - 1);
-		double inv_height = 1.0 / (static_cast<double>(raytrace_settings.image_height) - 1);
+		float inv_width = 1.0f / (static_cast<float>(raytrace_settings.image_width) - 1);
+		float inv_height = 1.0f / (static_cast<float>(raytrace_settings.image_height) - 1);
 		std::for_each(
 			std::execution::par,
 			data.pixels.begin(),
@@ -141,8 +141,8 @@ public:
 			[&pixel_colors, &world, &camera, render_settings,
 				inv_samples_per_pixel, inv_width, inv_height](raytrace_pixel& pixel)
 			{
-				const double u = (pixel.x + random::static_double.get()) * inv_width;
-				const double v = (pixel.y + random::static_double.get()) * inv_height;
+				const float u = (pixel.x + random::static_double.get()) * inv_width;
+				const float v = (pixel.y + random::static_double.get()) * inv_height;
 				pixel.color += ray_color_with_gradient_sky_attenuated(camera.compute_ray_to(u, v), world, render_settings, color::white(), color::black());
 				
 				// convert pixel.color into image-readable ascii pixel_colors
@@ -168,10 +168,10 @@ private:
 		while (true)
 		{
 			hit_info hit{&lambertian_material::default_material()};
-			if (!world.hit(raycast, 0.001, constants::infinity, hit))
+			if (!world.hit(raycast, 0.001f, constants::infinity, hit))
 			{
-				const double t = 0.5 * (raycast.direction.y() + 1.0);
-				return color(acc_emitted + mul(acc_attenuation, color(((1.0 - t) * settings.background_bottom_color + t * settings.background_top_color) * settings.background_strength)));
+				const float t = 0.5f * (raycast.direction.y() + 1.0f);
+				return color(acc_emitted + mul(acc_attenuation, color(((1.0f - t) * settings.background_bottom_color + t * settings.background_top_color) * settings.background_strength)));
 			}
 			
 			color attenuation;
